pyproject.toml
[tool.poetry]
name = "python-powerpoint-chart"
version = "0.1.0"
description = ""
authors = ["Sudeep Hazra <sudeephazra@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-pptx = "^0.6.23"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
